# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Backend for the Quantum Learning Laboratory.\n\nThe quantum engine computes, the authored evaluator grades, and the tutor\nteaches. Private rubrics and answer keys are never returned by any route in\nthis contract.",
    "title": "Quantum Learning Laboratory API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/course/{chapter_id}": {
      "get": {
        "operationId": "read_chapter_api_v1_course__chapter_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Chapter Api V1 Course  Chapter Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Chapter",
        "tags": [
          "content"
        ]
      }
    },
...
    "/api/v1/topics/{topic_id}": {
      "get": {
        "operationId": "read_topic_api_v1_topics__topic_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "topic_id",
            "required": true,
            "schema": {
              "title": "Topic Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Topic Api V1 Topics  Topic Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Topic",
        "tags": [
          "content"
        ]
      }
    },
...
  }
}